+++ docs/src/release_notes.rst
@@ -8,7 +8,7 @@ Release Notes
 
 0.8.20
 ------
-+ Fixed deletion in the batch module for TC instances < 5.7
++ Fixed deletion in the batch module for TC instances < 5.7.
 
 0.8.19
 ------